+++ scheduler/src/main/scala/mesosphere/mesos/Constraints.scala
@@ -1,0 +1,10 @@
+package mesosphere.mesos
+
+import org.apache.mesos.Protos.Attribute
+
+trait Placed {
+  def attributes: Seq[Attribute]
+  def hostname: Option[String]
+  def region: Option[String]
+  def zone: Option[String]
+}
